gsettings m4: Use --strict for checking
A while ago we allowed glib-compile-schemas to return a 'success' status
in the case that just one schema file contained errors.  Of course, this
is the exact opposite of what we want in the case that we are checking
schema validity at compile time.

Use the --strict flag for that case.

This closes #633115.