Bug 653841: Update README.win32 and VS README.txt's
This relates to my previous commit titled "add a script to generator
files for building" on behalf of Shixin Zeng.

Tell people about the availability of a python script to create the
necessary files for a Visual C++ build from a GIT checkout.

This is done with the courtesy of Shixin Zeng's python script which does
the job and eliminates the troubles of getting a suitable shell environment
to do the "make dist" job (which is especially not easy on Windows itself!)