gtester: Count inability to run a test in a binary as a failure
Previously, when a binary did not run a single test - for whatever
reason, from the binary not existing to the binary not using the gtester
framework - no failures were recorded. Now we record a non-successful
run of a binary that did not start any tests as a failure, too.

Note that this does not change the behavior of any binaries that exit
successfully or that report the start of any gtester test.