GParam: try to avoid further invalid uses
In an attempt to avoid some potential future abuses of the GParamSpec
API, qualify the 'name' field of the structure as 'const' and add a
comment noting that it is an interned string.

This is a theoretical API break, but it will only ever result in
warnings -- and even then, only if you were already doing something
questionable.

Clean up some of the warnings that were caused internally in gparam.c
from these changes.