Add GWakeup
GWakeup is a utility class to handle the cross-thread signalling needs
of GMainContext and GCancellable.  It may find some other users as well.

The desire here is to properly hide the implementation details in a
module which can be properly unit tested and used in GMainContext and
GCancellable without a rats nest of #ifdef.