gtester: small Windows fix
This code was unconditionally present in the gtester Makefile:

test-nonrecursive: ${TEST_PROGS}

On Windows, our testcases are compiled with a .exe suffix.  That means
that if we had 'foo' in TEST_PROGS, running "make check" would depend on
'foo' (not foo.exe) being compiled.

We could bring the EXEEXT in here to fix that up, but gtester doesn't
work on Windows at all, so better to just disable it in that case.